requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
Django==3.2.9
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
Werkzeug==2.0.2
yarl==1.7.2
